requirements.txt
boto3
mail-parser
pydantic
python-dotenv
supabase
pypinyin
requests
fastapi
uvicorn[standard]
playwright
beautifulsoup4
cryptography
python-multipart
aiohttp
asyncpg
sqlalchemy[asyncio]
alembic
pydantic-settings
redis[asyncio]
httpx
psutil